app/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
pyjwt>=2.10.1
bcrypt==4.1.3
motor==3.3.1
python-multipart>=0.0.9
Pillow>=10.2.0
numpy>=1.26.0
python-jose>=3.3.0
tzdata>=2024.2

typer>=0.9.0
razorpay



